python/requirements.txt
slack-bolt==1.18.0
python-gitlab==3.15.0
schedule==1.2.0
PyGithub==2.1.1
python-dotenv==1.0.0